move lsq functions from qat to trainable infra
